# docs/swagger.yaml
basePath: /
definitions:
  models.BPFProgram:
    properties:
      admin_status:
        description: Program admin status enabled or disabled
        type: string
      artifact:
        description: Artifact file name
        type: string
      cfg_version:
        description: Config version
        type: integer
      cmd_config:
        description: Program config providing command
        type: string
      cmd_start:
        description: Program start command
        type: string
      cmd_status:
        description: Program status command
        type: string
      cmd_stop:
        description: Program stop command
        type: string
      cmd_update:
        description: Program update config command
        type: string
      config_args:
        allOf:
        - $ref: '#/definitions/models.L3afDNFArgs'
        description: Map of arguments to config command
      config_file_path:
        description: Config file location
        type: string
      cpu:
        description: User program cpu limits
        type: integer
      ebpf_package_repo_url:
        description: Download url for Program
        type: string
      entry_function_name:
        description: BPF entry function name to load
        type: string
      id:
        description: Program id
        type: integer
      is_plugin:
        description: User program is plugin or not
        type: boolean
      map_args:
        description: Config BPF Map of arguments
        items:
          $ref: '#/definitions/models.L3afDMapArg'
        type: array
      map_name:
        description: BPF map to store next program fd
        type: string
      memory:
        description: User program memory limits
        type: integer
      monitor_maps:
        description: Metrics BPF maps
        items:
          $ref: '#/definitions/models.L3afDNFMetricsMap'
        type: array
      name:
        description: Name of the BPF program package
        type: string
      object_file:
        description: Object file contains BPF code
        type: string
      prog_type:
        description: Program type XDP or TC
        type: string
      rules:
        description: Config rules
        type: string
      rules_file:
        description: Config rules file name
        type: string
      seq_id:
        description: Sequence position in the chain
        type: integer
      start_args:
        allOf:
        - $ref: '#/definitions/models.L3afDNFArgs'
        description: Map of arguments to start command
      status_args:
        allOf:
        - $ref: '#/definitions/models.L3afDNFArgs'
        description: Map of arguments to status command
      stop_args:
        allOf:
        - $ref: '#/definitions/models.L3afDNFArgs'
        description: Map of arguments to stop command
      update_args:
        allOf:
        - $ref: '#/definitions/models.L3afDNFArgs'
        description: Map of arguments to update command
      user_program_daemon:
        description: User program daemon or not
        type: boolean
      version:
        description: Program version
        type: string
    type: object
  models.BPFProgramNames:
    properties:
      probes:
        description: names of the probe eBPF programs
        items:
          type: string
        type: array
      tc_egress:
        description: names of the TC egress eBPF programs
        items:
          type: string
        type: array
      tc_ingress:
        description: names of the TC ingress eBPF programs
        items:
          type: string
        type: array
      xdp_ingress:
        description: names of the XDP ingress eBPF programs
        items:
          type: string
        type: array
    type: object
  models.BPFPrograms:
    properties:
      probes:
        description: list of probe bpf programs
        items:
          $ref: '#/definitions/models.BPFProgram'
        type: array
      tc_egress:
        description: list of tc egress bpf programs
        items:
          $ref: '#/definitions/models.BPFProgram'
        type: array
      tc_ingress:
        description: list of tc ingress bpf programs
        items:
          $ref: '#/definitions/models.BPFProgram'
        type: array
      xdp_ingress:
        description: list of xdp ingress bpf programs
        items:
          $ref: '#/definitions/models.BPFProgram'
        type: array
    type: object
  models.KeyValue:
    properties:
      key:
        description: Key
        type: integer
      value:
        description: Value
        type: integer
    type: object
  models.L3afBPFProgramNames:
    properties:
      bpf_programs:
        allOf:
        - $ref: '#/definitions/models.BPFProgramNames'
        description: List of eBPF program names to remove
      host_name:
        description: Host name or pod name
        type: string
      iface:
        description: Interface name
        type: string
    type: object
  models.L3afBPFPrograms:
    properties:
      bpf_programs:
        allOf:
        - $ref: '#/definitions/models.BPFPrograms'
        description: List of bpf programs
      host_name:
        description: Host name or pod name
        type: string
      iface:
        description: Interface name
        type: string
    type: object
  models.L3afDMapArg:
    properties:
      args:
        description: BPF map arguments
        items:
          $ref: '#/definitions/models.KeyValue'
        type: array
      name:
        description: BPF map name
        type: string
    type: object
  models.L3afDNFArgs:
    additionalProperties: true
    type: object
  models.L3afDNFMetricsMap:
    properties:
      aggregator:
        description: Aggregation function names
        type: string
      key:
        description: Index of the bpf map
        type: integer
      name:
        description: BPF map name
        type: string
    type: object
info:
  contact: {}
  description: Configuration APIs to deploy and get the details of the eBPF Programs
    on the node
  title: L3AFD APIs
  version: "1.0"
paths:
  /l3af/configs/v1:
    get:
      consumes:
      - application/json
      description: Returns details of the configuration of eBPF Programs for all interfaces
        on a node
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Returns details of the configuration of eBPF Programs for all interfaces
        on a node
  /l3af/configs/v1/{iface}:
    get:
      consumes:
      - application/json
      description: Returns details of the configuration of eBPF Programs for a given
        interface
      parameters:
      - description: interface name
        in: path
        name: iface
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Returns details of the configuration of eBPF Programs for a given interface
  /l3af/configs/v1/add:
    post:
      consumes:
      - application/json
      description: Adds new eBPF Programs on node
      parameters:
      - description: BPF programs
        in: body
        name: cfgs
        required: true
        schema:
          items:
            $ref: '#/definitions/models.L3afBPFPrograms'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Adds new eBPF Programs on node
  /l3af/configs/v1/delete:
    post:
      consumes:
      - application/json
      description: Removes eBPF Programs on node
      parameters:
      - description: BPF program names
        in: body
        name: cfgs
        required: true
        schema:
          items:
            $ref: '#/definitions/models.L3afBPFProgramNames'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Removes eBPF Programs on node
  /l3af/configs/v1/restart:
    put:
      consumes:
      - application/json
      description: Store meta data about ebpf programs and exit
      parameters:
      - description: BPF programs
        in: body
        name: cfgs
        required: true
        schema:
          items:
            $ref: '#/definitions/models.L3afBPFPrograms'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Store meta data about ebpf programs and exit
  /l3af/configs/v1/update:
    post:
      consumes:
      - application/json
      description: Update eBPF Programs configuration
      parameters:
      - description: BPF programs
        in: body
        name: cfgs
        required: true
        schema:
          items:
            $ref: '#/definitions/models.L3afBPFPrograms'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update eBPF Programs configuration
swagger: "2.0"